Adding E2E TCP proxy support and test case to E2E dol

+++ nic/tools/run_e2e_tlsproxy_test.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python
+#!/usr/bin/python3
 
 import os
 import sys
@@ -8,6 +8,22 @@ import time
 import subprocess
 import threading
 
+paths = [
+    '/nic/gen/proto/',
+    '/nic/gen/proto/hal/',
+    '/nic'
+]
+ws_top = os.path.dirname(sys.argv[0]) + '/../../'
+ws_top = os.path.abspath(ws_top)
+for path in paths:
+    fullpath = ws_top + path
+    print("Adding Path: %s" % fullpath)
+    sys.path.insert(0, fullpath)
+
+import grpc
+import types_pb2
+import proxy_pb2
+
 from threading import Thread
 
 from subprocess import Popen, PIPE, call
@@ -16,19 +32,56 @@ nic_tools_dir = os.path.dirname(sys.argv[0])
 nic_dir = nic_tools_dir + "/.."
 nic_dir = os.path.abspath(nic_dir)
 
-print "NIC_DIR = ", nic_dir
+print("NIC_DIR = ", nic_dir)
 if nic_dir is None:
-    print "NIC dir is not set!"
+    print("NIC dir is not set!")
     sys.exit(1)
 
 pid = os.getpid()
 hntap_process = None
 tls_svr_process = None
+tcp_svr_process = None
 sock_stats = []
 
 hntap_log = nic_dir + "/hntap.log"
 tls_svr_log = nic_dir + "/tls_server.log"
+tcp_svr_log = nic_dir + "/tcp_server.log"
 tcp_clt_log = nic_dir + "/tcp_client.log"
+HalChannel = None
+
+def hal_init():
+    global HalChannel
+
+    if 'HAL_GRPC_PORT' in os.environ:
+        port = os.environ['HAL_GRPC_PORT']
+    else:
+        port = '50054'
+    print("Creating GRPC channel to HAL")
+    server = 'localhost:' + port
+    HalChannel = grpc.insecure_channel(server)
+    print("Waiting for HAL to be ready ...")
+    grpc.channel_ready_future(HalChannel).result()
+    print("Connected to HAL!")
+    return
+
+#def set_paths():
+
+
+def set_proxy_tls_bypass_mode(bypass_mode):
+    #set_paths()
+    hal_init()
+    stub = proxy_pb2.ProxyStub(HalChannel)
+    req_msg = proxy_pb2.ProxyGlobalCfgRequestMsg()
+    req_spec = req_msg.request.add()
+    req_spec.proxy_type = types_pb2.PROXY_TYPE_TLS
+    req_spec.bypass_mode = bypass_mode
+    resp_msg = stub.ProxyGlobalCfg(req_msg)
+    if resp_msg.api_status[0] != types_pb2.API_STATUS_OK:
+        print(" Failed to set TLS Proxy bypass mode: HAL API Status:%d" % (resp_msg.api_status[0]))
+        assert(0)
+        return
+    print("Set TLS Proxy bypass mode:%d" % bypass_mode)
+    return    
 
 def run_hntap():
     log = open(hntap_log, "w")
@@ -36,8 +89,8 @@ def run_hntap():
     p = Popen(cmd, stdout=log, stderr=log)
     global hntap_process
     hntap_process = p
-    print "* Starting Host-Network tapper, pid (" + str(p.pid) + ")"
-    print "    - Log file: " + hntap_log + "\n"
+    print("* Starting Host-Network tapper, pid (%s)" % str(p.pid))
+    print("    - Log file: " + hntap_log + "\n")
 
     log2 = open(hntap_log, "r")
     loop = 1
@@ -53,41 +106,57 @@ def run_hntap():
     return
 
 
-def run_tls_server():
+def run_tls_server(tcp_port):
     log = open(tls_svr_log, "w")
-    cmd = ['../bazel-bin/nic/proxy-e2etest/nic_proxy-e2etest_tls-server', str(80)]
+    cmd = ['../bazel-bin/nic/proxy-e2etest/nic_proxy-e2etest_tls-server', tcp_port ]
     p = Popen(cmd, stdout=log, stderr=log)
     global tls_svr_process
     tls_svr_process = p
-    print "* Starting TLS Server on port 80, pid (" + str(p.pid) + ")"
-    print "    - Log file: " + tls_svr_log + "\n"
+    print("* Starting TLS Server on port %s, pid (%s)" % (tcp_port, str(p.pid)))
+    print("    - Log file: " + tls_svr_log + "\n")
+    time.sleep(2)
+    return
+
+def run_tcp_server(tcp_port):
+    log = open(tcp_svr_log, "w")
+    cmd = ['../bazel-bin/nic/proxy-e2etest/nic_proxy-e2etest_tcp-server', tcp_port ]
+    p = Popen(cmd, stdout=log, stderr=log)
+    global tcp_svr_process
+    tcp_svr_process = p
+    print("* Starting TLS Server on port %s, pid (%s)" % (tcp_port, str(p.pid)))
+    print("    - Log file: " + tcp_svr_log + "\n")
     time.sleep(2)
     return
 
 
-def run_tcp_client():
-    log = open(tcp_clt_log, "w")
-    cmd = ['../bazel-bin/nic/proxy-e2etest/nic_proxy-e2etest_tcp-client', '-p', str(80), '-d', nic_dir + "/proxy-e2etest/hello-world", '-m', 'from-host']
+def run_tcp_client(tcp_port):
+    log = open(tcp_clt_log, "a")
+    cmd = ['../bazel-bin/nic/proxy-e2etest/nic_proxy-e2etest_tcp-client', '-p', tcp_port, '-d', nic_dir + "/proxy-e2etest/hello-world", '-m', 'from-host']
     p = Popen(cmd, stdout=log, stderr=log)
-    print "* Starting TCP Client on port 80, pid (" + str(p.pid) + ")"
-    print "    - Log file: " + tcp_clt_log + "\n"
+    print("* Starting TCP Client on port %s, pid (%s)" % (tcp_port, str(p.pid)))
+    print("    - Log file: " + tcp_clt_log + "\n")
     p.communicate()
     p.wait()
     return p.returncode
 
 def cleanup(keep_logs=True):
     os.kill(int(hntap_process.pid), 9)
-    os.kill(int(tls_svr_process.pid), 9)
+    if tls_svr_process:
+        os.kill(int(tls_svr_process.pid), 9)
+    if tcp_svr_process:
+        os.kill(int(tcp_svr_process.pid), 9)
 
     if keep_logs:
         return
 
     for log in hntap_log, tls_svr_log, tcp_clt_log:
         if os.path.isfile(log):
-            print "- " + log
+            print("- " + log)
             os.remove(log)
         else:
-            print "- " + log + " - not found"
+            print("- " + log + " - not found")
+            print("not found")
+
 
 def check_sockets(id, stop):
    global sock_stats
@@ -97,7 +166,7 @@ def check_sockets(id, stop):
      num_lines = 0
      est_socks = []
      for line in iter(p.stdout.readline, ''):
-         if "ESTABLISHED" in line:
+         if b"ESTABLISHED" in line:
              est_socks.append(line)
              num_lines += 1
      if (num_lines == 2):
@@ -118,57 +187,88 @@ def check_sockets(id, stop):
    sys.stdout.write("\n")
 
 def print_logs():
-    log = open(tcp_clt_log, "r")
-    print "TCP Client log:"
-    for line in log.readlines():
-       if "Client:" in line:
-           print "    " + line
-    log.close()
+    if os.path.isfile(tcp_clt_log):
+        log = open(tcp_clt_log, "r")
+        print("TCP Client log:")
+        for line in log.readlines():
+            if "Client:" in line:
+                print ("    " + line)
+        log.close()
 
-    print "TLS Server log:"
-    log = open(tls_svr_log, "r")
-    for line in log.readlines():
-       if "Server:" in line:
-           print "    " + line
-    log.close()
+    if os.path.isfile(tls_svr_log):
+        print("TLS Server log:")
+        log = open(tls_svr_log, "r")
+        for line in log.readlines():
+            if "Server:" in line:
+                print("    " + line)
+        log.close()
+
+    if os.path.isfile(tcp_svr_log):
+        print("TCP Server log:")
+        log = open(tcp_svr_log, "r")
+        for line in log.readlines():
+            if "Server:" in line:
+                print("    " + line)
+        log.close()
+
     if len(sock_stats) != 0:
-      print "Socket states:"
+      print("Socket states:")
       for line in sock_stats:
-          print "    " + line
+          print("    " + line)
 
 
-def main():
+def run_test(testnum, testname, tcp_port):
+    print("Test %d: Running E2E %s Proxy test, tcp-port %s\n" % (testnum, testname, tcp_port))
     start_time = time.time()
-    #os.chdir(nic_dir)
-    run_hntap()
-    run_tls_server()
 
-    stop_thread = False
-    t = Thread(target = check_sockets, args=(id, lambda: stop_thread))
-    t.start()
+    if (testname == "TCP"):
+        set_proxy_tls_bypass_mode(True)
+        run_tcp_server(tcp_port)
+    else:
+        set_proxy_tls_bypass_mode(False)
+        run_tls_server(tcp_port)
 
-    status = run_tcp_client()
-    if status != 0:
-        cleanup(keep_logs=True)
-        stop_thread = True
-        t.join()
-        print_logs()
-        print("\n- Total run time: %s seconds\n" % round(time.time() - start_time, 1))
-        print "Final Status = FAIL\n"
-        #We'll ignore the result to monitor stability of the test for a few days
-        status = 0
-        sys.exit(status)
-    
-    cleanup(keep_logs=True)
-    stop_thread = True
-    t.join()
+    status = run_tcp_client(tcp_port)
 
     print_logs()
-    print("\n- Total run time: %s seconds\n" % round(time.time() - start_time, 1))
-    print "Final Status = PASS\n"
+    print("\n- Test %d: run time: %s seconds\n" % (testnum, round(time.time() - start_time, 1)))
+    if status != 0:
+        print("Test %d: E2E %s Proxy Status = FAIL\n" % (testnum, testname))
+    else:
+        print("Test %d: E2E %s Proxy Status = PASS\n" % (testnum, testname))
+    return status
 
-# cleanup(keep_logs=False)
-    sys.exit(0)
+def main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--config-only", dest='configonly', action="store_true",
+                        help="Configuration only, no tests run")
+    args = parser.parse_args()
+
+    if args.configonly:
+        set_proxy_tls_bypass_mode(True)
+        run_hntap()
+        print("E2E TLS DOL: config-only, stopping..\n")
+        sys.exit(0)
+        
+    start_time = time.time()
+    log = open(tcp_clt_log, "w")
+    log.close()
+
+    run_hntap()
+
+    status = run_test(1, "TLS", str(80))
+    if status == 0:
+        time.sleep(5)
+        status = run_test(2, "TCP", str(81))
+
+    cleanup(keep_logs=True)
+
+    print("\n- Total run time: %s seconds\n" % round(time.time() - start_time, 1))
+    print("Final Status = %s\n" % ("PASS" if (status == 0) else "FAIL"))
+
+    #We'll ignore the result to monitor stability of the test for a few days
+    status = 0
+    sys.exit(status)
 
 if __name__ == "__main__":
     main()
